Add script to make bundle with Ubuntu.
This creates a bundle with Blender 2.76, MCell 3.3, and the development
branch of CellBlender. Once it's ready, this will be updated to point to
CellBlender 1.1 instead. The plotters, gamer, and other things still
need to be added, but this is a reasonable first step at making a Linux
bundle. This was made with Ubuntu 14.04, but it should work on other
distros with minor adjustments.