Fix how zip is created.
The zip was being created in such a way that it preserved the full file
path, such that, when you extracted it, it would create
/home/vagrant/bundle_cellblender/blender-2.77a-linux-glibc211-x86_64
instead of just blender-2.77a-linux-glibc211-x86_64.

This probably needs to be done for the 16.04 branch too. I thought I had
already made this change, but it must have never been pushed because I
can't find it anywhere in the commit history.