Address broken matplotlib-noqt recipe.
I was using a custom matplotlib recipe that didn't use qt, but it seems
to be broken, so I just reverted to that official version of matplotlib.
This bloats up the bundle by several hundred megs but is functional as
long as you force tkagg (i.e. matplotlib.use("TkAgg")).