Use new matplotlib recipe (w/o Qt).
I forked an existing matplotlib recipe and changed it so that it no
longer uses Qt. This brings the overall size of the cellblender build
back down to a much more reasonable level. This approach still needs to
be tested in Windows, but it at least seems to work for Linux.