Update Windows make CB bundle script.
I'm doing away with the custom matplotlib build, because it was just too
much trouble to maintain. Also, I'm using anaconda instead of miniconda.
The compressed build is under 400 MB, which isn't terrible. It's
unfortunate, because we could really make do with just tk. There's no
strong incentive to use qt and it's so huge, but we'll live with it now,
since it will save us a lot of maintenance.

I still have to update some hard coded paths and update the
instructions, which I'll do in a future commit.